+++ pkg/local_object_storage/blobstor/internal/storagetest/delete.go
@@ -1,4 +1,4 @@
-package blobstortest
+package storagetest
 
 import (
 	"testing"